// src/api/types.ts
export interface ValueObject {
}

export interface Address extends ValueObject {
    line1: string;
    line2: string | undefined;
    city: string;
    state: string | undefined;
    postalCode: string;
    countryAlpha2: string;
}